move "models" to `db/` folder
