+++ db/Answer.js
@@ -1,4 +1,4 @@
-const db = require("../db");
+const db = require("./");
 
 function Answer() {
   return db("Answers");
