+++ db/Group.js
@@ -1,38 +1,37 @@
-const db = require("../db");
+const db = require("./");
 
-function Questionnaire() {
-  return db("Questionnaires");
+function Group() {
+  return db("Groups");
 }
 
 module.exports.findAll = function findAll(where = {}) {
-  return Questionnaire()
+  return Group()
     .where(where)
     .select();
 };
 
-
 module.exports.findById = function findById(id) {
-  return Questionnaire()
+  return Group()
     .where("id", parseInt(id, 10))
     .first();
 };
 
 module.exports.update = function update(id, updates) {
-  return Questionnaire()
-    .where("id", parseInt(id, 10))
+  return Group()
+    .where({ "id" : parseInt(id, 10) })
     .update(updates)
     .returning("*");
 };
 
 module.exports.create = function create(obj) {
-  return Questionnaire()
+  return Group()
     .insert(obj)
     .returning("*");
 };
 
 module.exports.destroy = function destroy(id) {
-  return Questionnaire()
-    .where("id", parseInt(id, 10))
+  return Group()
+    .where({ "id" : parseInt(id, 10) })
     .delete()
     .returning("*");
 }
