+++ db/Page.js
@@ -1,4 +1,4 @@
-const db = require("../db");
+const db = require("./");
 
 function Page() {
   return db("Pages");
