+++ db/Question.js
@@ -1,4 +1,4 @@
-const db = require("../db");
+const db = require("./index");
 
 function Question() {
   return db("Pages");
