+++ db/Questionnaire.js
@@ -1,37 +1,38 @@
-const db = require("../db");
+const db = require("./");
 
-function Group() {
-  return db("Groups");
+function Questionnaire() {
+  return db("Questionnaires");
 }
 
 module.exports.findAll = function findAll(where = {}) {
-  return Group()
+  return Questionnaire()
     .where(where)
     .select();
 };
 
+
 module.exports.findById = function findById(id) {
-  return Group()
+  return Questionnaire()
     .where("id", parseInt(id, 10))
     .first();
 };
 
 module.exports.update = function update(id, updates) {
-  return Group()
-    .where({ "id" : parseInt(id, 10) })
+  return Questionnaire()
+    .where("id", parseInt(id, 10))
     .update(updates)
     .returning("*");
 };
 
 module.exports.create = function create(obj) {
-  return Group()
+  return Questionnaire()
     .insert(obj)
     .returning("*");
 };
 
 module.exports.destroy = function destroy(id) {
-  return Group()
-    .where({ "id" : parseInt(id, 10) })
+  return Questionnaire()
+    .where("id", parseInt(id, 10))
     .delete()
     .returning("*");
 }
